requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.4.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
mlflow==2.19.0
matplotlib==3.10.0
seaborn==0.13.2
shap==0.45.0
requests==2.31.0
jupyter==1.1.1
openai
python-dotenv==1.0.1
xgboost
pre-commit